Fixed some syntax warnings (== instead of ===)
